Create a new voice package to support external connections
