+++ heartbeat.go
@@ -1,7 +1,6 @@
 package harmony
 
 import (
-	"encoding/binary"
 	"sync/atomic"
 	"time"
 
@@ -32,54 +31,4 @@ func (c *Client) sendHeartbeatPayload() error {
 	}
 	atomic.StoreInt64(&c.lastHeartbeatSend, time.Now().UnixNano())
 	return c.sendPayload(gatewayOpcodeHeartbeat, sequence)
-}
-
-// heartbeat periodically sends a heartbeat payload to the voice server.
-func (vc *VoiceConnection) heartbeat(every time.Duration) {
-	vc.logger.Debug("starting voice connection heartbeater")
-	defer vc.logger.Debug("stopped voice connection heartbeater")
-
-	heartbeat.Run(
-		&vc.wg,
-		vc.stop,
-		vc.error,
-		every,
-		vc.sendHeartbeatPayload,
-		&vc.lastHeartbeatACK,
-	)
-}
-
-// sendHeartbeatPayload sends a single heartbeat payload
-// to the voice server containing a nonce.
-func (vc *VoiceConnection) sendHeartbeatPayload() error {
-	return vc.sendPayload(voiceOpcodeHeartbeat, time.Now().Unix())
-}
-
-// udpHeartbeat periodically sends a UDP heartbeat packet to the voice server.
-func (vc *VoiceConnection) udpHeartbeat(every time.Duration) {
-	vc.logger.Debug("starting UDP heartbeater")
-	defer vc.logger.Debug("stopped UDP heartbeater")
-
-	heartbeat.RunUDP(
-		&vc.wg,
-		vc.stop,
-		vc.error,
-		time.Second*5,
-		vc.sendUDPHeartbeat,
-		&vc.lastUDPHeartbeatACK,
-	)
-}
-
-// sendUDPHeartbeat sends a single UDP heartbeat packet and increments the sequence number.
-func (vc *VoiceConnection) sendUDPHeartbeat() error {
-	packet := make([]byte, 8)
-
-	// Load and increment the UDP sequence atomically,
-	// but send the value before the increment.
-	binary.LittleEndian.PutUint64(packet, atomic.AddUint64(&vc.udpHeartbeatSequence, 1)-1)
-	if _, err := vc.udpConn.Write(packet); err != nil {
-		return err
-	}
-
-	return nil
 }
