+++ listen.go
@@ -21,21 +21,3 @@ func (c *Client) listen() {
 func (c *Client) recvPayloads(ch chan<- *payload.Payload) {
 	payload.RecvAll(&c.wg, ch, c.error, c.recvPayload)
 }
-
-// listen listens for payloads sent by the voice server.
-func (vc *VoiceConnection) listen() {
-	vc.logger.Debug("starting voice connection event listener")
-	defer vc.logger.Debug("stopped voice connection event listener")
-
-	payload.Listen(
-		&vc.wg,
-		vc.stop,
-		vc.error,
-		vc.recvPayloads,
-		vc.handleEvent,
-	)
-}
-
-func (vc *VoiceConnection) recvPayloads(ch chan<- *payload.Payload) {
-	payload.RecvAll(&vc.wg, ch, vc.error, vc.recvPayload)
-}
